+++ Sources/DuplicationDetector/Algorithms/ExactCloneDetector.swift
@@ -195,7 +195,8 @@ public struct ExactCloneDetector: Sendable {
 
         // Roll through remaining windows
         let maxStartIndex = tokens.count - minimumTokens
-        for i in 1...maxStartIndex where maxStartIndex >= 1 {
+        guard maxStartIndex >= 1 else { return windows }
+        for i in 1...maxStartIndex {
             let outgoing = tokenTexts[i - 1]
             let incoming = tokenTexts[i + minimumTokens - 1]
             hash = rollingHash.roll(hash: hash, outgoing: outgoing, incoming: incoming)